Create Data pattern analyser.py
(d) A data pattern analyser is to be built that can analyse a sequence and detect
and count up the number of occurrences of three consecutive integers where the
third integer is the product of the first and second integers. Integers can be positive
or negative. The analyser needs to stop counting and terminate immediately if it
reaches an element in the sequence with value 0.
So:
• Sequence (1, -1, 0) would return 0 (termination condition)
• Sequence (1, -2, -2) would return 1
• Sequence (1, 2, 3, 6) would return 1
• Sequence (1, 2, 2, 4) would return 2 (the sequences (1, 2, 2) and (2, 2, 4))
• Sequence (1, 2, 2, 0, 2, 4, 8, 32) would return 1 (because it stops counting
and terminates immediately after reaching the element with value 0).